Sistema em andamento: login funcional: sim cadastro de cliente: sim listagem de cliente:sim edição de clientes: sim exclusão de cliente: sim geração de relatorio em pdf: sim envio de email: sim logout funcional: sim

+++ src/main/java/model/CadastroClientes.java
@@ -285,7 +285,7 @@ public class CadastroClientes extends JFrame {
 	public void salvarNovoCliente() {
 		DaoCliente dao = new DaoCliente();
 		String buscaUltimoId = dao.buscaId();
-		int id = (Integer.parseInt(buscaUltimoId) + 1);
+		long id = (Integer.parseInt(buscaUltimoId) + 1);
 		String nomeCliente = nome.getText();
 		String telefoneCliente = telefone.getText();
 		String emailCliente = email.getText();
